backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.9
python-docx==1.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
aiofiles==23.2.0
python-jose[cryptography]==3.3.0
